moved some stuff into constants and utilities files + updated imports
