+++ src/applications/ivc-champva/10-10D/tests/unit/helpers/helpers.unit.spec.js
@@ -5,8 +5,7 @@ import {
   getAgeInYears,
   makeHumanReadable,
 } from '../../../../shared/utilities';
-import { sponsorWording } from '../../../helpers/wordingCustomization';
-import { isInRange } from '../../../helpers/utilities';
+import { isInRange, sponsorWording } from '../../../helpers/utilities';
 import { getTopLevelFormData } from '../../../components/Applicant/applicantFileUpload';
 import ApplicantField from '../../../../shared/components/applicantLists/ApplicantField';
 import { testComponentRender } from '../../../../shared/tests/pages/pageTests.spec';
